dashboard/requirements.txt
# ==========================================================
#  Dashboard (Streamlit) — Local viewer for Mongo/JSON
# ==========================================================

streamlit>=1.38.0
plotly>=5.24.1
pandas>=2.2.2
numpy>=1.26.4
pymongo==4.8.0
python-dotenv>=1.0.1

# Optional (for notebook / visualization experiments)
matplotlib>=3.9.0
jupyterlab>=4.2.3